Made the code more efficient
Made the code more efficient, added a progress bar from 0 to 100, added 3 new arguments
The arguments:
-d [DIR]
Specifies which extensions must be in which directory. Works only in combination with -e
-e [EXTENSION]
Specifies which extensions must be in which directory. Works only in combination with -d
--help
Returns the help page and exit